requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
langchain-text-splitters
langchain-groq
fastapi
uvicorn
langserve
sse_starlette